Moved enums into their own namespace
