+++ src/main/java/com/everneth/emi/commands/bot/ApplyCommand.java
@@ -1,15 +1,13 @@
 package com.everneth.emi.commands.bot;
 
 import com.everneth.emi.EMI;
-import com.everneth.emi.models.ConfigMessage;
+import com.everneth.emi.models.enums.ConfigMessage;
 import com.everneth.emi.models.EMIPlayer;
 import com.everneth.emi.models.WhitelistApp;
 import com.everneth.emi.services.WhitelistAppService;
 import com.jagrosh.jdautilities.command.SlashCommand;
 import net.dv8tion.jda.api.entities.Role;
 import com.jagrosh.jdautilities.command.SlashCommandEvent;
-import net.dv8tion.jda.api.exceptions.ErrorHandler;
-import net.dv8tion.jda.api.requests.ErrorResponse;
 
 public class ApplyCommand extends SlashCommand {
     public ApplyCommand()
